requirements.txt
requests
mysql-connector-python
pandas
python-dotenv
schedule
sqlalchemy
pytz
logging
pyyaml
aiohttp
asyncio
